# internal/pkg/openapi/models/сonfirm.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"bytes"
	"context"
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Сonfirm сonfirm
//
// swagger:model Сonfirm
type Сonfirm struct {

	// confirmation сode
	// Required: true
	ConfirmationСode *string `json:"confirmationСode"`

	// transaction ID
	// Required: true
	TransactionID *string `json:"transactionID"`
}

// UnmarshalJSON unmarshals this object while disallowing additional properties from JSON
func (m *Сonfirm) UnmarshalJSON(data []byte) error {
	var props struct {

		// confirmation сode
		// Required: true
		ConfirmationСode *string `json:"confirmationСode"`

		// transaction ID
		// Required: true
		TransactionID *string `json:"transactionID"`
	}

	dec := json.NewDecoder(bytes.NewReader(data))
	dec.DisallowUnknownFields()
	if err := dec.Decode(&props); err != nil {
		return err
	}

	m.ConfirmationСode = props.ConfirmationСode
	m.TransactionID = props.TransactionID
	return nil
}

// Validate validates this сonfirm
func (m *Сonfirm) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateConfirmationСode(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateTransactionID(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Сonfirm) validateConfirmationСode(formats strfmt.Registry) error {

	if err := validate.Required("confirmationСode", "body", m.ConfirmationСode); err != nil {
		return err
	}

	return nil
}

func (m *Сonfirm) validateTransactionID(formats strfmt.Registry) error {

	if err := validate.Required("transactionID", "body", m.TransactionID); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this сonfirm based on context it is used
func (m *Сonfirm) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Сonfirm) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Сonfirm) UnmarshalBinary(b []byte) error {
	var res Сonfirm
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}