# internal/pkg/openapi/swagger.yaml
swagger: "2.0"
info:
  description: Bank transfer system
  title: scam-master
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http

paths:
  /health-check:
    get:
      tags:
        - Standard
      operationId: HealthCheck
      summary: Health check
      description: Checking the server health status.
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              ok:
                type: boolean

  /init-transaction:
    post:
      tags:
        - Transactions
      operationId: Init
      summary: Init transaction
      description: Initialize funds transfer from card to card.
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/Transaction"
      responses:
        204:
          description: OK
        400: {$ref: '#/responses/BadRequest'}
        default: {$ref: '#/responses/InternalError'}
  
  /confirm-transaction:
    post:
      tags:
        - Transactions
      operationId: Confirm
      summary: Confirm transaction
      description: Confirm the funds transfer with a confirmation code.
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/Сonfirm'
      responses:
        204:
          description: OK
        400: {$ref: '#/responses/BadRequest'}
        404: {$ref: '#/responses/NotFound'}
        default: {$ref: '#/responses/InternalError'}

responses:
  BadRequest:
    description: Bad request
    schema:
      $ref: '#/definitions/Error'
  Forbidden:
    description: Forbidden
    schema:
      $ref: '#/definitions/Error'
  NotFound:
    description: Not found
    schema:
      $ref: '#/definitions/Error'
  InternalError:
    description: Internal error
    schema:
      $ref: '#/definitions/Error'

definitions:
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        description: Either same as HTTP Status Code OR >= 600
        type: integer
        format: int32
      message:
        type: string
  
  Transaction:
    type: object
    properties:
      transactionID:
        type: string
      senderCardNumber:
        type: string
        pattern: '^\d{16}$'
        example: '1111111111111111'
      validity:
        type: string
        pattern: '^\d{4}$'
        example: '0101'
      cvc:
        type: string
        pattern: '^\d{3}$'
        example: '111'
      recipientCardNumber:
        type: string
        pattern: '^\d{16}$'
        example: '1111111111111111'
      bankGateway:
        description: The bank through which the funds will be transferred
        type: string
        enum: ['tinkoff']
      senderBank:
        description: Sender's card bank
        type: string
        enum: ['tinkoff']
      amount:
        description: Amount in kopecks (RUB * 10^2)
        type: integer
        format: int64
        minimum: 1
    required:
      - transactionID
      - senderCardNumber
      - validity
      - cvc
      - recipientCardNumber
      - bankGateway
      - senderBank
      - amount
  
  Сonfirm:
    type: object
    properties:
      transactionID:
        type: string
      confirmationСode:
        type: string
    required:
      - transactionID
      - confirmationСode